06_challenge/mcp/requirements.txt
fastapi
uvicorn
praw
python-dotenv
pydantic
fastapi-mcp
tavily-python
